Added a documentation for channel, compiler, constant modules with `echo` + `ping_pong` examples.

+++ constant/include/constant/constant.h
@@ -1,5 +1,6 @@
 #pragma once
 
+//
 // @TODO:
 //
 // use https://en.cppreference.com/w/cpp/thread/hardware_destructive_interference_size
@@ -12,5 +13,5 @@
 
 namespace ihft::constant
 {
-    constexpr unsigned long CPU_CACHE_LINE_SIZE = 64;
+    constexpr unsigned CPU_CACHE_LINE_SIZE = 64;
 }
